Tests accès API OpenRouteService passants
