+++ src/test/java/fr/diginamic/VroomVroomCar/service/OpenRouteServiceTest.java
@@ -1,24 +1,15 @@
 package fr.diginamic.VroomVroomCar.service;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.web.client.RestTemplate;
-
-import java.util.HashMap;
-import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
@@ -28,7 +19,6 @@ import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 class OpenRouteServiceTest {
-
     @Mock
     private RestTemplate restTemplate;
 
@@ -38,45 +28,12 @@ class OpenRouteServiceTest {
     @InjectMocks
     private OpenRouteService openRouteService;
 
+
     @BeforeEach
     void setUp() {
         ReflectionTestUtils.setField(openRouteService, "API_KEY", "clé-fictive");
         ReflectionTestUtils.setField(openRouteService, "objectMapper", new ObjectMapper());
     }
-
-    @Value("${openroute.api.key}")
-    private String API_KEY;
-
-    @Test
-    public void testApiKeyValidity() {
-        String testUrl = "https://api.openrouteservice.org/geocode/search?api_key=" + API_KEY + "&text=Paris";
-
-        RestTemplate restTemplate = new RestTemplate();
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.set("Accept", "application/json, application/geo+json, application/gpx+xml, img/png; charset=utf-8");
-        headers.set("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64)");
-
-        HttpEntity<String> entity = new HttpEntity<>(headers);
-
-        try {
-            ResponseEntity<String> response = restTemplate.exchange(testUrl, HttpMethod.GET, entity, String.class);
-
-            System.out.println("Status Code: " + response.getStatusCode());
-            System.out.println("Response Body: " + response.getBody());
-
-            if (response.getStatusCode().is2xxSuccessful()) {
-                System.out.println("✅ Clé API valide !");
-            } else {
-                System.out.println("❌ Clé API invalide ou problème de configuration");
-            }
-
-        } catch (Exception e) {
-            System.out.println("❌ Erreur lors du test: " + e.getMessage());
-            e.printStackTrace();
-        }
-    }
-
 
     @Test
     void testGetCoordinatesFromAddress() throws Exception {
